o/devicestate, asserts: fixed nosecboot tests and added missed review fix
